chore: Moves `CartItem` to fs/components
